fix: make mobile menu reusable
